reorganize __main__ into cli module
